FIX: Remove union types with null
In previous commit of this file, nullable fields are typeds as union of
string and null type. Now they are declare as optional fields in the
interface.

EDIT: Forgot to save the file and to mention that lastActive field is
removed from StrictUserInput as it was only for testing Date custom
scalar