CHORE: Add custom IMapUserType and use dataloader in mapper
Created IMapUserType interface to match return type of mapper function
mapUser, in which the returned object has async function interests which
returns IFieldType objects on demand.

It's a good practise to assign a async function which to property which
returns list of objects. In this way, we are not loading the whole
object, instead we are provinding a function, on called, returns the
desired output.