FEAT: Changes types and set bio to nullable property
Model:
Set birthDate default value to null, since birthDate is not
required to be set anymore in the new schema.

In validators, assign model to constant variable User with type hints,
so it'd be easy to read which it is.

Types:
Remove NativeDate and use Date constructor instead to specify types to
properties.

Set bio and birthDate a nullable property.

Only export IUserBaseDocument and IUserDocument, other exports are
removed