FIX: Return first element of groupedUsers[id]
groupedUsers stores list of objects satisfy the given predicate. Since user id is unique, there must be only one element in matched ids. And also return entity must be list of objects, and not list of list of objects. Therefore returning first element is the right way to do this.